# refactor: moving tests to classes to avoid repeated fixtures execution