--- a/doc/changelog.d/3523.added.md
+++ b/doc/changelog.d/3523.added.md
@@ -1,0 +1,1 @@
+refactor: moving tests to classes to avoid repeatingly execute the fixtures.